Use dot key for normal mode
With spacebar, there was an annoying interaction with VLC. Which I think was exacerbated by the fact that VLC was triggering pause on the down event, rather than up.

What we ought to do really is stop the `last_key` events from getting through at all, by grabbing, relaying events to a virtual device (with uinput), and not immediately letting events through while RIGHT_ALT is held. But we've had issues with calling from the right thread when we tried grabbing (as opposed to using `xinput`).