Grab evdev devices
Note that unlike the Rust version we're using evdev grabbing instead of xinput. I think the reason we did things differently before was because there were issues surrounding which thread did the grabbing. But everything _seems_ to work here. And if we had to, I think we could configure Streamly to use a single OS thread anyway.

NB. even though we don't do anything with the error in `RemoveKeyboard` in `runAction`, it's still useful because it appears in the logs when `show`-ing `RemoveKeyboard d e`.