Simplify hifi power toggling
No longer any need to ever use ctrl-p. We always toggle the plug, reading its current status in order to know whether to turn on or off. We could alternatively use `was_on` from the response and store it, since in practice we never trigger from other sources. This would avoid making two HTTP calls, but I'm not sure it's worth the complexity since there's no noticeable lag.

We also _could_ make this a single action, since the sub-actions aren't used elsewhere. This would be a little simpler. But some day I'd like to maybe use them for automating turning the TV on (although it's not clear how best that might work unless we always want it on when the HiFi is on - it would be better to turn the HiFi on when turning TV on, but we can't get the current power level of the TV, only blindly toggle it).

NB. `ToggleHifiPlug` is now unused, but there's no real harm in keeping it around.