Add keybindings for launching programs
These are the only GUI programs that usually get used on this system. We don't want to keep them all running all of the time due to resource constraint, but we do want it to be easy to open them.

This was actually prompted by the fact that Kodi tends to freeze when left open for a while. This has always been an occasional issue, but has happened consistently since upgrading to a RPi 5.