Migrate to new hi-fi
This has no remote control, so we're more limited in options for controlling volume etc. The intention for now is to largely use software controls, both system and Spotifyd-level, and keep the analog output on the unit at maximum. This isn't ideal in principal for fidelity, but actually output is crackly at other volumes for now anyway, possibly pending cleaning the voume knob contacts.