Revert "Communicate with Spotifyd directly for transferring playback"
This reverts commit e997eafcd6ebd2f759ebf8ecce76d817afd814be.

Unfortunately, when playback wasn't already on the Pi, the command failed due to an unrecognised target. It is possible that there's some potential workaround, since this seems like it'd be a universal problem. But anyway, since we transfer playback a lot less often than pausing, resuming, or especially modifying volume, there's much less of a downside to just using the Spotify HTTP API here.